+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.2b2 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Registry records are now correctly removed when uninstalling the package.
+  [hvelarde]
+
 - Package is compatible with Plone 4.2 again.
   [hvelarde]